api-gateway/quiz_service/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.11
openai==1.6.1
kafka-python==2.0.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlalchemy==2.0.25
boto3==1.34.20
